+++ apps/web/src/app/admin-console/organizations/members/components/bulk/bulk-enable-sm-dialog.component.ts
@@ -15,7 +15,7 @@ export type BulkEnableSecretsManagerDialogData = {
 };
 
 @Component({
-  templateUrl: `bulk-enable-sm.component.html`,
+  templateUrl: `bulk-enable-sm-dialog.component.html`,
 })
 export class BulkEnableSecretsManagerDialogComponent implements OnInit {
   protected dataSource = new TableDataSource<OrganizationUserView>();
